Abstract logic out of app.rs
